Separated out Py3 specific tests to avoid running under py2 (c) Copyright IBM Corp. 2015
